archive/mlx_training_distributed/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "training-mlx"
version = "0.1.0"
description = "Training MLX - Distributed Training System for MLX Models"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "George Dekker", email = "george@mlx.ai"}
]
keywords = ["machine learning", "distributed training", "fine-tuning", "mlx", "lora"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core MLX (optional - works without it)
    # "mlx>=0.22.0",
    # "mlx-lm>=0.21.0",
    # API Framework (Port 8200)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    # Configuration and CLI
    "pyyaml>=6.0",
    "click>=8.1.0",
    "jsonschema>=4.0.0",
    # Utilities
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    # Logging
    "rich>=13.0.0",
    # Distributed Integration (optional)
    # "mlx-distributed>=0.1.0",
    "mlx>=0.27.1",
    "mlx-lm>=0.26.2",
    "pandas>=2.3.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "aiohttp>=3.12.15",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.22.0",
    "mlx-lm>=0.21.0",
]
training = [
    "datasets>=3.0.0",
    "transformers>=4.48.0",
    "tokenizers>=0.21.0",
    "numpy>=1.24.0",
    "wandb>=0.19.0",
    "tensorboard>=2.18.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
training-mlx = "training_mlx.cli.main:cli"

[project.urls]
Homepage = "https://github.com/ml-explore/training-mlx"
Documentation = "https://training-mlx.readthedocs.io/"
Repository = "https://github.com/ml-explore/training-mlx"
Issues = "https://github.com/ml-explore/training-mlx/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.ruff]
target-version = "py313"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # function calls in defaults
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--verbose"
